requirements.txt
flask==3.1.0
werkzeug==3.1.3
nltk==3.9.1
spacy==3.8.5
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.0.2
python-docx==1.1.2
PyPDF2==3.0.1
sentence-transformers==4.1.0
gunicorn==23.0.0
boto3==1.38.3
